Refactor Identifiable base class to use long instead of int for ID types across various resource models and tests

+++ test/DiscoveryTests/PrivateResourceDefinition.cs
@@ -6,4 +6,4 @@ namespace DiscoveryTests;
 
 [UsedImplicitly(ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature)]
 public sealed class PrivateResourceDefinition(IResourceGraph resourceGraph)
-    : JsonApiResourceDefinition<PrivateResource, int>(resourceGraph);
+    : JsonApiResourceDefinition<PrivateResource, long>(resourceGraph);
