+++ test/JsonApiDotNetCoreTests/IntegrationTests/Logging/AuditEntry.cs
@@ -6,7 +6,7 @@ namespace JsonApiDotNetCoreTests.IntegrationTests.Logging;
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 [Resource(ControllerNamespace = "JsonApiDotNetCoreTests.IntegrationTests.Logging")]
-public sealed class AuditEntry : Identifiable<int>
+public sealed class AuditEntry : Identifiable<long>
 {
     [Attr]
     public string UserName { get; set; } = null!;
