+++ test/JsonApiDotNetCoreTests/IntegrationTests/NonJsonApiControllers/KnownResource.cs
@@ -6,7 +6,7 @@ namespace JsonApiDotNetCoreTests.IntegrationTests.NonJsonApiControllers;
 
 [UsedImplicitly(ImplicitUseTargetFlags.Members)]
 [Resource(ControllerNamespace = "JsonApiDotNetCoreTests.IntegrationTests.NonJsonApiControllers")]
-public sealed class KnownResource : Identifiable<int>
+public sealed class KnownResource : Identifiable<long>
 {
     [Attr]
     public string? Value { get; set; }
